Move OpenBSD-specific BPF functionality into its own file.
On Linux, BPF behaves differently enough for the bpf_event/bpf_input split.
bpf_input now accepts whole packets only and the size of the read buffer is
controlled on an OS basis too.

A couple of functions had to be de-static'd, but it shouldn't affect binary
size or performance very drastically.

Signed-off-by: Martin Pelikan <pelikan@storkhole.cz>