refactor and clarify request/renewing checks
When L3 IP src != BOOTP ciaddr, we don't treat that as an error any more.
This situation happens when a laptop has wired and wireless interfaces on
and the unicast renewal for one goes through the other.  As per RFC 2131,
we only enforce ciaddr and chaddr mapping but still log a note otherwise.